requirements.txt
streamlit
panda
numpy 
scikit-learn
matplotlib
seaborn
